requirements-v2.0.txt
# Core Application
streamlit
pandas
numpy # Removed specific version for now, let pip resolve
python-dateutil
python-dotenv==1.0.1 # Keep specific version
requests

# Database
aiosqlite==0.21.0         # async SQLite support
# faiss-cpu==1.10.0       # Can be commented out if not used for metadata similarity yet
SQLAlchemy==2.0.39 # Keep specific version

# GenAI & Embeddings
# openai==1.68.2          # Commented out as we focus on Gemini now
# cohere==5.14.0          # Commented out
langchain-google-genai   # ADDED for Gemini support
google-generativeai      # Often a dependency, good to list
# sentence-transformers==4.0.2 # Commented out if not used for embeddings yet
# huggingface-hub==0.30.1 # Commented out if not used
# transformers            # Commented out if not used
tiktoken # Often needed by LangChain tokenizers
# tokenizers==0.21.1      # Commented out if not used directly

# LangChain & LangGraph
langchain
# langchain-cohere        # Commented out
langchain-community
langchain-core
# langchain-milvus        # Commented out
# langchain-ollama        # Commented out
# langchain-openai        # Commented out
langchain-text-splitters
langchainhub
# langfuse                # Commented out if not using LangFuse specifically
langgraph                # REQUIRED
# langgraph-api           # Usually part of langgraph or langserve install
# langgraph-checkpoint    # Needed for persistence, maybe later
langgraph-cli            # REQUIRED
# langgraph-prebuilt      # Optional prebuilt graphs
# langgraph-sdk           # Usually part of langgraph install
langsmith                # Recommended for tracing

# LangServe (for LangGraph CLI Server)
langserve                # REQUIRED

# Data Science, ML and DL (Keep for potential future DSA use)
scikit-learn # Removed specific version
scipy        # Removed specific version
statsmodels  # Removed specific version
# keras==3.9.2 # Commented out if not used now
# xgboost==3.0.0 # Commented out if not used now
joblib       # Removed specific version
torch        # Keep if planning DL later

# Data Visualization
matplotlib # Removed specific version
plotly     # REQUIRED
seaborn    # Removed specific version

# Python
six # Removed specific version
pydantic-settings # Removed specific version
pydantic         # REQUIRED (Often v1 needed by LangChain, pip should handle)

# Utilities & Helpers
black # Removed specific version
cloudpickle # Removed specific version
click     # Removed specific version
pypdf
openpyxl # Removed specific version
tqdm

# Web Frameworks (for LangServe)
# Flask==3.1.0 # Commented out, LangServe uses FastAPI
fastapi    # REQUIRED
uvicorn    # REQUIRED (often needs [standard] extras)
# beautifulsoup4 # Commented out if not used
# boto3==1.36.26 # Commented out if not used
# botocore==1.36.26 # Commented out if not used
Jinja2 # Removed specific version